requirements.txt
streamlit==1.42.0
python-docx==1.1.2
python-dotenv==1.0.1
pandas==2.2.3
google-cloud-bigquery==3.29.0
crewai==0.100.1
